56 users viewmodel (#58)
* created userRepository

* created user viewmodel

* edited signup

user gets added to database when a signup occurs

---------

+++ app/src/main/java/com/example/appdev2_assignment2/CRUD/UserCRUD.kt
@@ -88,7 +88,8 @@ class UserRepositoryFirestore(var db: FirebaseFirestore):UserRepository {
 fun convertDocumentToUser(document: DocumentSnapshot): AppUser{
     var name = document.getString("name") ?: ""
     var email = document.getString("email") ?: ""
-    var age = document.getLong("age")?.toByte() ?: 0
+    var age = document.getLong("age")?.toInt() ?: 0
+
     var picture = document.getLong("picture")?.toInt() ?: 0
 
     return AppUser(email, name, age, picture)
