+++ app/src/main/java/com/example/appdev2_assignment2/User.kt
@@ -3,4 +3,5 @@ package com.example.appdev2_assignment2
 data class User (var email:String)
 
 
-class AppUser(var email: String, var name: String, var age: Byte, var picture: Int)
+class AppUser(var email: String, var name: String, var age: Int, var picture: Int)
+
